+++ src/shared.ts
@@ -1,0 +1,9 @@
+export const formatBytes = (bytes: number, decimals = 2): string => {
+  // https://stackoverflow.com/a/18650828
+  if (bytes == 0) return '0 Bytes';
+  const k = 1024;
+  const dm = decimals === undefined ? 2 : decimals;
+  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + sizes[i];
+};
