+++ src/ui/components/breadcrumb.ts
@@ -1,0 +1,10 @@
+import Vue from 'vue';
+
+export default Vue.extend({
+  props: {
+    tab: {
+      type: String,
+      required: true,
+    },
+  },
+});
