+++ src/ui/components/glossary/link.ts
@@ -1,0 +1,14 @@
+import Vue from 'vue';
+
+export default Vue.extend({
+  props: {
+    to: {
+      type: String,
+      required: true,
+    },
+    text: {
+      type: String,
+      default: '',
+    },
+  },
+});
