+++ src/ui/components/isolation/settings.vue
@@ -1,33 +1,5 @@
 <script>
-export default {
-  props: {
-    action: {
-      type: String,
-      required: true,
-    },
-    glossary: {
-      type: String,
-      default: '',
-    },
-    glossaryLabel: {
-      type: String,
-      default: '',
-    },
-    label: {
-      type: String,
-      default: '',
-    },
-    perdomain: {
-      type: Boolean,
-      default: false,
-    },
-  },
-  watch: {
-    action(newAction) {
-      this.$emit('update:action', newAction);
-    },
-  },
-};
+export { default } from './settings.ts';
 </script>
 
 <template>
