+++ src/ui/components/root.vue
@@ -1,0 +1,7 @@
+<script>
+export { default } from './root.ts';
+</script>
+
+<template>
+  <component :is="App" :app="app" />
+</template>
