+++ src/ui/components/statistics.vue
@@ -1,64 +1,5 @@
 <script>
-export default {
-  props: {
-    app: {
-      type: Object,
-      required: true,
-    },
-  },
-  data() {
-    return {
-      preferences: this.app.preferences,
-      permissions: this.app.permissions,
-      statistics: this.app.storage.statistics,
-      popup: this.app.popup,
-    };
-  },
-  async mounted() {
-    $('#statistics .ui.checkbox').checkbox();
-
-    if (!this.popup) {
-      $('#deletesHistoryStatisticsField').popup({
-        html: `
-          <div style="width:500px;">
-          The overall statistics include all Temporary Containers already<br>
-          This will show and collect separate statistics about how many "Deletes History<br>
-          Temporary Container" plus cookies and URLs with them got deleted.</div>
-        `,
-        inline: true,
-        position: 'bottom left',
-      });
-    }
-  },
-  methods: {
-    async resetStatistics() {
-      if (
-        !window.confirm(`
-        Reset statistics?
-      `)
-      ) {
-        return;
-      }
-
-      await browser.runtime.sendMessage({
-        method: 'resetStatistics',
-      });
-
-      this.$root.$emit('initialize', {
-        showMessage: 'Statistics have been reset.',
-      });
-    },
-    formatBytes(bytes, decimals) {
-      // https://stackoverflow.com/a/18650828
-      if (bytes == 0) return '0 Bytes';
-      const k = 1024;
-      const dm = decimals === undefined ? 2 : decimals;
-      const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-      const i = Math.floor(Math.log(bytes) / Math.log(k));
-      return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + sizes[i];
-    },
-  },
-};
+export { default } from './statistics.ts';
 </script>
 
 <template>
