+++ src/ui/popup.ts
@@ -1,5 +1,3 @@
-import App from './popup.vue';
-
-import('./root').then(root => {
-  root.default(App, { popup: true });
-});
+import App from './components/popup.vue';
+import root from './root';
+root(App, { popup: true });
